requirements.txt
asgiref==3.7.2
Django==4.2.2
django-jsonfield==1.4.1
django-unixtimestampfield==0.5.0
djangorestframework==3.14.0
psycopg2==2.9.6
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
